Codes/requirements.txt
beautifulsoup4==4.10.0
Brotli==1.0.9
bs4==0.0.1
camelot-py==0.10.1
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
cryptography==3.4.8
dash==2.0.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
DateTime==4.3
decorator==5.1.0
Elixir==0.7.1
et-xmlfile==1.1.0
Flask==2.0.1
Flask-Compress==1.10.1
ghostscript==0.7
gunicorn==20.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.21.2
opencv-python==4.5.3.56
opencv-python-headless==4.5.3.56
openpyxl==3.0.8
pandas==1.3.3
patsy==0.5.1
pbr==5.6.0
pdfminer.six==20201018
pdftopng==0.2.3
plotly==5.3.1
plotly-express==0.4.1
pycparser==2.20
PyPDF2==1.26.0
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
scipy==1.7.1
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.2.1
SQLAlchemy==0.7.10
sqlalchemy-migrate==0.11.0
sqlparse==0.4.2
statsmodels==0.12.2
tabulate==0.8.9
Tempita==0.5.2
tenacity==8.0.1
tk==0.1.0
urllib3==1.26.6
Werkzeug==2.0.1
xlrd==0.7.1
xlwt==0.7.2
zope.interface==5.4.0
